+++ daemon/cmd/daemon/main.go
@@ -2,7 +2,7 @@ package main
 
 import "log"
 
-var Version = "v2.2.6"
+var Version = "v2.2.8"
 
 func main() {
 	if err := runDaemon(parseDaemonOptions()); err != nil {